requirements.txt
scikit-learn
mojimoji
pyyaml
tqdm
gensim